+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Use .net 6.0
 - Use .net 7.0
 - Add prefix option
+- Add rename command by image meta data
 
 ## 1.0.0 - 2020-10-26
 - Initial implementation